Moved all intermediary methods to TranslAg.
