+++ src/main/java/jacamo/rest/RestImplAg.java
@@ -1,28 +1,14 @@
 package jacamo.rest;
 
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.FileOutputStream;
-import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringReader;
 import java.io.StringWriter;
 import java.net.URI;
-import java.text.SimpleDateFormat;
 import java.util.ArrayList;
-import java.util.Date;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import java.util.TreeMap;
-import java.util.concurrent.Executor;
-import java.util.concurrent.Executors;
-import java.util.concurrent.locks.Condition;
-import java.util.concurrent.locks.Lock;
-import java.util.concurrent.locks.ReentrantLock;
-import java.util.logging.LogRecord;
-import java.util.logging.StreamHandler;
 
 import javax.inject.Singleton;
 import javax.ws.rs.Consumes;
@@ -43,37 +29,18 @@ import org.glassfish.jersey.media.multipart.FormDataContentDisposition;
 import org.glassfish.jersey.media.multipart.FormDataParam;
 import org.w3c.dom.Document;
 
-import com.google.common.reflect.ClassPath;
-import com.google.common.reflect.ClassPath.ClassInfo;
 import com.google.gson.Gson;
 
-import cartago.ArtifactId;
-import cartago.ArtifactInfo;
-import cartago.CartagoException;
-import cartago.CartagoService;
-import cartago.WorkspaceId;
-import jaca.CAgentArch;
 import jason.ReceiverNotFoundException;
-import jason.architecture.AgArch;
 import jason.asSemantics.Agent;
-import jason.asSemantics.CircumstanceListener;
-import jason.asSemantics.IntendedMeans;
-import jason.asSemantics.Intention;
-import jason.asSemantics.Option;
-import jason.asSemantics.TransitionSystem;
 import jason.asSemantics.Unifier;
 import jason.asSyntax.ASSyntax;
 import jason.asSyntax.Literal;
-import jason.asSyntax.Plan;
-import jason.asSyntax.PlanBody;
 import jason.asSyntax.PlanLibrary;
-import jason.asSyntax.Trigger;
-import jason.asSyntax.Trigger.TEType;
 import jason.asSyntax.VarTerm;
 import jason.asSyntax.parser.ParseException;
 import jason.infra.centralised.BaseCentralisedMAS;
 import jason.infra.centralised.CentralisedAgArch;
-import jason.stdlib.print;
 
 /**
  * Agent's REST compile class
@@ -86,7 +53,6 @@ import jason.stdlib.print;
 @Path("/agents")
 public class RestImplAg extends AbstractBinder {
 
-    Map<String, StringBuilder> agLog = new HashMap<>();
     TranslAg tAg = new TranslAg();
     Gson gson = new Gson();
 
@@ -95,8 +61,6 @@ public class RestImplAg extends AbstractBinder {
         bind(new RestImplAg()).to(RestImplAg.class);
     }
     
-    Executor executor = Executors.newFixedThreadPool(4);
-
     /**
      * Produces JSON containing the list of existing agents Example: ["ag1","ag2"]
      * 
@@ -106,20 +70,6 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAgentsJSON() {
         return Response.ok().entity(gson.toJson(tAg.getAgents())).header("Access-Control-Allow-Origin", "*").build();
-    }
-
-    /**
-     * Return agent object by agent's name
-     * 
-     * @param agName name of the agent
-     * @return Agent object
-     */
-    private Agent getAgent(String agName) {
-        CentralisedAgArch cag = BaseCentralisedMAS.getRunner().getAg(agName);
-        if (cag != null)
-            return cag.getTS().getAg();
-        else
-            return null;
     }
 
     /**
@@ -136,38 +86,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.TEXT_PLAIN)
     public Response createNewAgent(@PathParam("agentname") String agName) {
         try {
-            String givenName = BaseCentralisedMAS.getRunner().getRuntimeServices().createAgent(agName, null, null, null,
-                    null, null, null);
-            BaseCentralisedMAS.getRunner().getRuntimeServices().startAgent(givenName);
-            // set some source for the agent
-            Agent ag = getAgent(givenName);
-
-            try {
-
-                File f = new File("src/agt/" + givenName + ".asl");
-                if (!f.exists()) {
-                    f.createNewFile();
-                    FileOutputStream outputFile = new FileOutputStream(f, false);
-                    StringBuilder stringBuilder = new StringBuilder();
-                    stringBuilder.append("//Agent created automatically\n\n");
-                    stringBuilder.append("!start.\n\n");
-                    stringBuilder.append("+!start <- .print(\"Hi\").\n\n");
-                    stringBuilder.append("{ include(\"$jacamoJar/templates/common-cartago.asl\") }\n");
-                    stringBuilder.append("{ include(\"$jacamoJar/templates/common-moise.asl\") }\n");
-                    stringBuilder.append(
-                            "// uncomment the include below to have an agent compliant with its organisation\n");
-                    stringBuilder.append("//{ include(\"$moiseJar/asl/org-obedient.asl\") }");
-                    byte[] bytes = stringBuilder.toString().getBytes();
-                    outputFile.write(bytes);
-                    outputFile.close();
-                }
-
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-            ag.load(new FileInputStream("src/agt/" + givenName + ".asl"), givenName + ".asl");
-            // ag.setASLSrc("no-inicial.asl");
-            createAgLog(givenName, ag);
+            String givenName = tAg.createAgent(agName);
 
             return Response.created(new URI("/agents/" + givenName)).build();
         } catch (Exception e) {
@@ -213,7 +132,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAgentStatusJSON(@PathParam("agentname") String agName) {
         try {
-            Map<String, Object> props = getAgent(agName).getTS().getUserAgArch().getStatus();
+            Map<String, Object> props = tAg.getAgent(agName).getTS().getUserAgArch().getStatus();
             return Response.ok(gson.toJson(props)).build();
         } catch (Exception e) {
             e.printStackTrace();
@@ -257,7 +176,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.APPLICATION_XML)
     public Document getAgentMindXml(@PathParam("agentname") String agName) {
         try {
-            Agent ag = getAgent(agName);
+            Agent ag = tAg.getAgent(agName);
             if (ag != null)
                 return ag.getAgState();
             else
@@ -280,7 +199,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.APPLICATION_JSON)
     public Response getAgentBBJSON(@PathParam("agentname") String agName) {
         try {
-            Agent ag = getAgent(agName);
+            Agent ag = tAg.getAgent(agName);
             List<String> bbs = new ArrayList<>();
             for (Literal l : ag.getBB()) {
                 bbs.add(l.toString());
@@ -309,7 +228,7 @@ public class RestImplAg extends AbstractBinder {
             @DefaultValue("all") @QueryParam("label") String label) {
         StringWriter so = new StringWriter();
         try {
-            Agent ag = getAgent(agName);
+            Agent ag = tAg.getAgent(agName);
             if (ag != null) {
                 PlanLibrary pl = ag.getPL();
                 if (label.equals("all"))
@@ -347,7 +266,7 @@ public class RestImplAg extends AbstractBinder {
             @FormDataParam("file") InputStream uploadedInputStream,
             @FormDataParam("file") FormDataContentDisposition fileDetail) {
         try {
-            Agent ag = getAgent(agName);
+            Agent ag = tAg.getAgent(agName);
             if (ag != null) {
                 ag.parseAS(new StringReader(plans), "RestAPI");
 
@@ -380,7 +299,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.TEXT_PLAIN)
     public Response loadPlans(@PathParam("agentname") String agName, String plans) {
         try {
-            Agent ag = getAgent(agName);
+            Agent ag = tAg.getAgent(agName);
             if (ag == null) {
                 return Response.status(500, "Receiver '" + agName + "' not found").build();
             }
@@ -393,59 +312,26 @@ public class RestImplAg extends AbstractBinder {
     }
 
     /**
-     * Get list of internal actions for an agent
+     * Return a TEXT PLAIN of available internal action, external actions and
+     * commands for the given agent Example:
+     * "['.desire','.drop_desire','.drop_all_desires']"
      * 
-     * @return List of internal actions
+     * @param agName Name of the agent
+     * @return HTTP 200 Response (ok status) or 500 Internal Server Error in case of
+     *         error (based on https://tools.ietf.org/html/rfc7231#section-6.6.1)
      */
-    public void getIASuggestions(Map<String,String> cmds) {
+    @Path("/{agentname}/code")
+    @GET
+    @Produces(MediaType.APPLICATION_JSON)
+    public Response getCodeCompletionSuggestions(@PathParam("agentname") String agName) {
+        Map<String,String> commands = new HashMap<>();
         try {
-            ClassPath classPath = ClassPath.from(print.class.getClassLoader());
-            Set<ClassInfo> allClasses = classPath.getTopLevelClassesRecursive("jason.stdlib");
-
-            allClasses.forEach(a -> {
-                try {
-                    Class<?> c = a.load();
-                    if (c.isAnnotationPresent(jason.stdlib.Manual.class)) {
-                        // add full predicate provided by @Manual
-                        jason.stdlib.Manual annotation = (jason.stdlib.Manual) c
-                                .getAnnotation(jason.stdlib.Manual.class);
-                        cmds.put(annotation.literal(), annotation.hint().replaceAll("\"", "`").replaceAll("'", "`"));
-					} else {
-						// add just the functor of the internal action
-						cmds.put("." + a.getSimpleName(), "");
-					}
-
-				} catch (Exception e) {
-					e.printStackTrace();
-				}
-			});
-
-		} catch (Exception e) {
-			e.printStackTrace();
-		}
-	}
-
-	/**
-	 * Return a TEXT PLAIN of available internal action, external actions and
-	 * commands for the given agent Example:
-	 * "['.desire','.drop_desire','.drop_all_desires']"
-	 * 
-	 * @param agName Name of the agent
-	 * @return HTTP 200 Response (ok status) or 500 Internal Server Error in case of
-	 *         error (based on https://tools.ietf.org/html/rfc7231#section-6.6.1)
-	 */
-	@Path("/{agentname}/code")
-	@GET
-	@Produces(MediaType.APPLICATION_JSON)
-	public Response getCodeCompletionSuggestions(@PathParam("agentname") String agName) {
-		Map<String,String> commands = new HashMap<>();
-		try {
             // get internal actions
-			getPlansSuggestions(agName, commands);
+            tAg.getPlansSuggestions(agName, commands);
             // get internal actions
-            getIASuggestions(commands);
+            tAg.getIASuggestions(commands);
             // get external actions
-            getEASuggestions(agName, commands);
+            tAg.getEASuggestions(agName, commands);
 
             Gson json = new Gson();
             Map<String,String> sortedCmds = new TreeMap<>(commands);
@@ -455,96 +341,6 @@ public class RestImplAg extends AbstractBinder {
         }
 
         return Response.status(500, "Server Internal Error! Could not get code completion suggestions.").build();
-    }
-
-	private void getPlansSuggestions(String agName, Map<String, String> commands) {
-		try {
-			// get agent's plans
-            Agent ag = getAgent(agName);
-            if (ag != null) {
-                PlanLibrary pl = ag.getPL();
-                for (Plan plan : pl.getPlans()) {
-                    
-                    // do not add plans that comes from jar files (usually system's plans)
-                    if (plan.getSource().startsWith("jar:file") || plan.getSource().equals("kqmlPlans.asl"))
-                        continue;
-
-                    // add namespace when it is not default
-                    String ns = "";
-                    if (!plan.getNS().equals(Literal.DefaultNS)) {
-                        ns = plan.getNS().toString() + "::";
-                    }
-
-                    String terms = "";
-                    if (plan.getTrigger().getLiteral().getArity() > 0) {
-                        for (int i = 0; i < plan.getTrigger().getLiteral().getArity(); i++) {
-                            if (i == 0)
-                                terms = "(";
-                            terms += plan.getTrigger().getLiteral().getTerm(i).toString();
-                            if (i < plan.getTrigger().getLiteral().getArity() - 1)
-                                terms += ", ";
-                            else
-                                terms += ")";
-                        }
-                    }
-
-                    // when it is a goal or test goal, do not add operator
-                    if ((plan.getTrigger().getType() == TEType.achieve)
-                            || (plan.getTrigger().getType() == TEType.test)) {
-
-
-                        commands.put(ns + plan.getTrigger().getType().toString()
-                        + plan.getTrigger().getLiteral().getFunctor() + terms, "");
-
-                    }
-                    // when it is belief, do not add type which is anyway empty
-                    else if (plan.getTrigger().getType() == TEType.belief) {
-                        commands.put(ns + plan.getTrigger().getOperator().toString()
-                                + plan.getTrigger().getLiteral().getFunctor() + terms, "");
-                    }
-                }
-            }
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-    }
-
-    private void getEASuggestions(String agName, Map<String, String> commands) throws CartagoException {
-
-        try {
-            Agent ag = getAgent(agName);
-            // get external actions (from focused artifacts)
-            CAgentArch cartagoAgArch = getCartagoArch(ag);
-            for (WorkspaceId wid : cartagoAgArch.getSession().getJoinedWorkspaces()) {
-                String wksName = wid.getName();
-                for (ArtifactId aid : CartagoService.getController(wksName).getCurrentArtifacts()) {
-
-                    // operations
-                    ArtifactInfo info = CartagoService.getController(wksName).getArtifactInfo(aid.getName());
-
-                    info.getObservers().forEach(y -> {
-                        if (y.getAgentId().getAgentName().equals(agName)) {
-                            info.getOperations().forEach(z -> {
-                                String params = "";
-                                for (int i = 0; i < z.getOp().getNumParameters(); i++) {
-                                    if (i == 0) params = "(";
-                                    params += "arg" + i;
-                                    if (i == z.getOp().getNumParameters() - 1)
-                                        params += ")";
-                                    else
-                                        params += ", ";
-                                }
-
-                                commands.put(z.getOp().getName() + params, "");
-                            });
-                        }
-                    });
-
-                }
-            }
-        } catch (CartagoException e) {
-            e.printStackTrace();
-        }
     }
 
     /**
@@ -561,18 +357,18 @@ public class RestImplAg extends AbstractBinder {
     @Consumes(MediaType.APPLICATION_FORM_URLENCODED)
     @Produces(MediaType.APPLICATION_JSON)
     public Response runCmdPost(@FormParam("c") String cmd, @PathParam("agentname") String agName) {
-        Agent ag = getAgent(agName);
+        Agent ag = tAg.getAgent(agName);
         if (ag == null) {
             return Response.status(500, "Receiver '" + agName + "' not found").build();
         }
         try {
-            createAgLog(agName, ag);
+            tAg.createAgLog(agName, ag);
             
             cmd = cmd.trim();
             if (cmd.endsWith(".")) cmd = cmd.substring(0, cmd.length() - 1);
 
-            Unifier u = execCmd(ag, ASSyntax.parsePlanBody(cmd));
-            addAgLog(agName, "Command " + cmd + ": " + u);
+            Unifier u = tAg.execCmd(ag, ASSyntax.parsePlanBody(cmd));
+            tAg.addAgLog(agName, "Command " + cmd + ": " + u);
 
             Map<String,String> um = new HashMap<String, String>();
             for (VarTerm v: u) {
@@ -599,7 +395,7 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.TEXT_PLAIN)
     public Response getLogOutput(@PathParam("agentname") String agName) {
         try {
-            StringBuilder o = agLog.get(agName);
+            StringBuilder o = tAg.agLog.get(agName);
             if (o != null) {
                 return Response.ok(o.toString()).build();
             }
@@ -623,138 +419,13 @@ public class RestImplAg extends AbstractBinder {
     @Produces(MediaType.TEXT_PLAIN)
     public Response delLogOutput(@PathParam("agentname") String agName) {
         try {
-            agLog.put(agName, new StringBuilder());
+            tAg.agLog.put(agName, new StringBuilder());
 
             return Response.ok().build();
         } catch (Exception e) {
             e.printStackTrace();
         }
         return Response.status(500).build();
-    }
-
-    /**
-     * Send a command to an agent
-     * 
-     * @param agName name of the agent
-     * @param sCmd   command to be executed
-     * @return Status message
-     * @throws ParseException 
-     */
-    Unifier execCmd(Agent ag, PlanBody lCmd) throws ParseException {
-        Trigger te = ASSyntax.parseTrigger("+!run_repl_expr");
-        Intention i = new Intention();
-        IntendedMeans im = new IntendedMeans(new Option(new Plan(null, te, null, lCmd), new Unifier()), te);
-        i.push(im);
-
-        Lock lock = new ReentrantLock();
-        Condition goalFinished  = lock.newCondition();
-        executor.execute( () -> {
-                /*GoalListener gl = new GoalListener() {                 
-                    public void goalSuspended(Trigger goal, String reason) {}
-                    public void goalStarted(Event goal) {}
-                    public void goalResumed(Trigger goal) {}
-                    public void goalFinished(Trigger goal, FinishStates result) {
-                        System.out.println("finished!");
-                        if (goal.equals(te)) {
-                            // finished
-                            //if (result.equals(FinishStates.achieved)) {
-                            //}                           
-                            try {
-                                lock.lock();
-                                goalFinished.signalAll();
-                            } finally {
-                                lock.unlock();
-                            }
-                        }                       
-                    }
-                    public void goalFailed(Trigger goal) {
-                        if (goal.equals(te)) {
-                            try {
-                                lock.lock();
-                                goalFinished.signalAll();
-                            } finally {
-                                lock.unlock();
-                            }
-                        }                                                   
-                    }
-                };*/
-                CircumstanceListener cl = new CircumstanceListener() {
-                    public void intentionDropped(Intention ci) {
-                        System.out.println("*finished!"+ci);
-                        if (ci.equals(i)) {
-                            try {
-                                lock.lock();
-                                goalFinished.signalAll();
-                            } finally {
-                                lock.unlock();
-                            }
-                        }
-                    };
-
-                };
-                TransitionSystem ts = ag.getTS();
-                try {
-                    lock.lock();
-                    //ts.addGoalListener(gl);
-                    ts.getC().addEventListener(cl);
-                    ts.getC().addRunningIntention(i);
-                    ts.getUserAgArch().wake();
-                    goalFinished.await();
-                    //ts.removeGoalListener(gl);
-                    ts.getC().removeEventListener(cl);
-                } catch (InterruptedException e) {                          
-                } finally {
-                    lock.unlock();
-                }
-                System.out.println("fim thread");
-        });
-        try {
-            lock.lock();
-            goalFinished.await();
-            
-            return im.getUnif();
-        } catch (InterruptedException e) {
-        } finally {
-            lock.unlock();
-        }
-        return null;
-    }
-
-    /**
-     * Creates a log area for an agent
-     * 
-     * @param agName agent name
-     * @param ag     agent object
-     */
-    protected void createAgLog(String agName, Agent ag) {
-        // adds a log for the agent
-        if (agLog.get(agName) == null) {
-            agLog.put(agName, new StringBuilder());
-            ag.getTS().getLogger().addHandler(new StreamHandler() {
-                @Override
-                public void publish(LogRecord l) {
-                    addAgLog(agName, l.getMessage());
-                }
-            });
-        }
-    }
-
-    /**
-     * Add a message to the agent log.
-     * 
-     * @param agName agent name
-     * @param msg    message to be added
-     */
-    protected void addAgLog(String agName, String msg) {
-        StringBuilder o = agLog.get(agName);
-        if (o == null) {
-            o = new StringBuilder();
-            agLog.put(agName, o);
-        } else {
-            o.append("\n");
-        }
-        String dt = new SimpleDateFormat("dd-MM-yy HH:mm:ss").format(new Date());
-        o.append("[" + dt + "] " + msg);
     }
 
     /**
@@ -803,22 +474,5 @@ public class RestImplAg extends AbstractBinder {
         }
 
         return Response.status(500).build();
-    }
-
-    /**
-     * Get agent's CArtAgO architecture
-     * 
-     * @param ag Agent object
-     * @return agent's CArtAgO architecture
-     */
-    protected CAgentArch getCartagoArch(Agent ag) {
-        AgArch arch = ag.getTS().getUserAgArch().getFirstAgArch();
-        while (arch != null) {
-            if (arch instanceof CAgentArch) {
-                return (CAgentArch) arch;
-            }
-            arch = arch.getNextAgArch();
-        }
-        return null;
     }
 }
